requirements.txt
fastapi==0.128.0
mysql-connector-python==9.5.0
pandas==3.0.0
pydantic==2.12.5
python-multipart==0.0.21
uvicorn==0.40.0
